feat(identifier-type): change mapping for 010 sub-fields
- Move entity creation logic to migration class

Closes MODSOURMAN-1127
