--- mod-source-record-manager-server/src/main/java/org/folio/services/migration/helper/FieldMappingHelper.java
@@ -1,61 +1,0 @@
-package org.folio.services.migration.helper;
-
-import io.vertx.core.json.JsonArray;
-import io.vertx.core.json.JsonObject;
-
-public class FieldMappingHelper {
-  private static final String SUBFIELD_A = "a";
-  private static final String SUBFIELD_Z = "z";
-  private static final String IDENTIFIER_TYPE_ID = "identifiers.identifierTypeId";
-  private static final String DESCRIPTION_TYPE_ID = "Identifier Type for ";
-  private static final String IDENTIFIER_VALUE = "identifiers.value";
-  private static final String DESCRIPTION_VALUE = "Library of Congress Control Number";
-  private static final String CANCELLED = "Cancelled ";
-  private static final String LCCN = "LCCN";
-
-
-  private FieldMappingHelper() {
-    throw new IllegalStateException("Utility class");
-  }
-
-  /**
-   * Create entities array of 010 field of Authority mapping rule
-   *
-   * @return entities JsonArray
-   */
-  public static JsonArray get010FieldEntityJsonArray() {
-    return JsonArray.of(
-      getEntityJsonObject(IDENTIFIER_TYPE_ID, DESCRIPTION_TYPE_ID + LCCN, SUBFIELD_A),
-      getEntityJsonObject(IDENTIFIER_VALUE, DESCRIPTION_VALUE, SUBFIELD_A),
-      getEntityJsonObject(IDENTIFIER_TYPE_ID, DESCRIPTION_TYPE_ID + CANCELLED + LCCN, SUBFIELD_Z),
-      getEntityJsonObject(IDENTIFIER_VALUE, CANCELLED + DESCRIPTION_VALUE, SUBFIELD_Z)
-    );
-  }
-
-  private static JsonObject getEntityJsonObject(String target, String description, String subfield) {
-    return JsonObject.of(
-      "target", target,
-      "description", description,
-      "subfield", JsonArray.of(subfield),
-      "rules", getRules(subfield.equals(SUBFIELD_A), target.equals(IDENTIFIER_TYPE_ID))
-    );
-  }
-
-  private static JsonArray getRules(boolean isSubfieldA, boolean isTypeIdRules) {
-    String prefix = isSubfieldA ? "" : CANCELLED;
-    return JsonArray.of(JsonObject.of("conditions", isTypeIdRules ? getTypeIdConditions(prefix) : getValueConditions()));
-  }
-
-  private static JsonArray getTypeIdConditions(String prefix) {
-    var conditionObject = JsonObject.of(
-      "type", "set_identifier_type_id_by_name",
-      "parameter", JsonObject.of("name", prefix + LCCN)
-    );
-    return JsonArray.of(conditionObject);
-  }
-
-  private static JsonArray getValueConditions() {
-    var type = JsonObject.of("type", "trim");
-    return JsonArray.of(type);
-  }
-}
